+++ Changelog.md
@@ -19,6 +19,7 @@
 - Lighting: Fixed minor shadow visual glitches occurring in some fields
 - Renderer: Fix black color in some field maps (`spipe2` for example) ( https://github.com/julianxhokaxhiu/FFNx/pull/587 )
 - Voice: Enable tutorial voice acting
+- Widescreen: Added experimental support for 16:10 aspect ratio
 
 ## FF8